Revert "Allow signposts to take advantage of deferred string substitution"
This reverts commits ce80830cf346c6fa74667f6b0b891082076e6121 and
ddc7c79e5e8e65fc3090fee576327a5bcaf09af5.

As explained in the original commit message, this didn't have the
intended effect of improving the common LLDB use case, but still
provided a marginal improvement for the places where LLDB creates a
scoped time with a string literal.

The reason for the revert is that this change pulls in the os/signpost.h
header in Signposts.h. The former transitively includes loader.h, which
contains a series of macro defines that conflict with MachO.h. There are
ways to work around that, but Adrian and I concluded that  none of them
are worth the trade-off in complicating Signposts.h even further.